feat: bump Starknet version to 0.13.4 (#169)
Starknet version ties heavily to the minimum Cairo version the network supports - deploying a Cairo contract that was compiled with unsupported compiler version may result in an error. 

Assuming that Dojo wants to use Cairo compiler version >=2.10, Katana should at least be using Starknet version **0.13.4** as Cairo 2.10 introduces new changes that are only supported starting from **0.13.4**.

We maintain the default Starknet version used in the genesis block - this is to ensure we change the computed genesis block hash.

---

The issue for this particular case, currently Katana is failing with a `Syscall gas cost must be greater than base syscall gas cost` error when executing a contract that uses the `get_class_hash_at_syscall`. The syscall was first introduced in Cairo **2.10.0**, and thus requires at least Starknet **0.13.4** as mentioned in the [release notes](https://community.starknet.io/t/cairo-v2-10-0-is-out/115362).