refactor(log): remove debug logs (#175)
Removes 'dev' logs which are basically just `debug`-level RPC server logs that would be implicitly enabled when Katana is used with `--dev` flag.

The logs have been found to be rather noisy as it will log out the method name of all RPC requests. The logs, however, can be manually enabled by adding `server=debug` to the `RUST_LOG` environment variable. So, it's not exactly a big deal for it to be removed.