refactor(contracts): feature contract types (#191)
The purpose of this refactor is mainly to avoid the hardcoding of class hashes by introducing a standardized API for defining and accessing feature contracts. 

Now, when adding a new feature contract, simply use the new `contract!` [procedural macro](https://doc.rust-lang.org/reference/procedural-macros.html) and several constant values are generated i.e., class hash, casm hash, and the class itself.

Example with the current set of feature contracts:

https://github.com/dojoengine/katana/blob/da25d47975da32333e7fec30576eb0f24ce96a3f/crates/contracts/src/lib.rs#L4-L7

The proc macro will generate three constants:

```rust
// Returns the class hash of the default account class
Account::HASH

// Returns the casm hash of the default account class
Account::HASH

// Get the ContractClass instance of the default account class
Account::CLASS
```

The old implementation in the `constant.rs` file is still being kept because replacing it with `katana-contracts` will introduce a cyclic dependency issue in `katana-primitives`. It should be noted that those are now deprecated and code that requires those constants should use the ones defined in `katana-contracts`. All usages outside of the primitive crate have already been updated to use the new constants.

https://github.com/dojoengine/katana/blob/49faff088a34650f6c59acf08f62a6fb9349f4a7/crates/primitives/src/genesis/constant.rs#L64-L81

Until we move the genesis stuff out of the `katana-primitives`, only then we can delete the `constant.rs` file.

This PR also removes the contract artifacts from git and requires that the artifacts be built first using `make build-contracts`. This is to ensure reproducible builds of the contract.