chore(utils): remove dependency on `dojo-utils` (#213)
The motivation to remove the reliance on `dojo-utils` is so to avoid having version conflicts when bumping shared dependencies (primarily the `starknet` crate). 

In some cases, when we're attempting to bump `starknet` crate version on Katana, it'd conflict with the version of `starknet` required by `dojo-utils`. Even though the stuff we're importing from `dojo-utils` doesn't rely on `starknet` at all (this is the case with the `dojo_utils::signal::wait_signals` function).