feat(rpc): update Starknet RPC to 0.9.0-rc.2 (#210)
An update to the Starknet JSON-RPC spec to version [0.9.0-rc.2](https://github.com/starkware-libs/starknet-specs/releases/tag/v0.9.0-rc.2).

Currently, the supported version of the Starknet JSON-RPC spec in Katana is 0.8.0. The Starknet network has planned on upgrading their RPC version to the new [0.9.0](https://community.starknet.io/t/sn-0-14-0-pre-release-notes/115618#p-2359352-rpc-09-5). It's already the version used on Starknet testnet now and will soon hit mainnet on [18 Aug](https://docs.starknet.io/resources/version-notes/#starknet_v0_14_0_aug_18_25).

As of writing this, the `starknet` crate has yet to publish a version on `crates.io` with the stable v0.9.0 (i.e., not one of the `rc`s) support and this is primarily because the stable version of the spec was just released very recently (2 days ago).