ci(release): allow specify custom version (#220)
Right now when running the `release-dispatch.yml` workflow, we can only choose from a set of options for the `version_type` input. This is fine if we have an ideal and linear history where every releases are made against the `main` branch, but sometime we break this rule and create a release at a commit that doesn't belong to the `main` branch (e.g., https://github.com/dojoengine/katana/releases/tag/v1.7.0-alpha.0). 

Due to my own negligence, I should've made the git history of the releases clean - the commit of each git tag (ie release version) and its predecessor can be tracked within the same branch. But now I've decided to abandon the branch of which the `v1.7.0-alpha.0` tag was made from (ie branch [`1.7.0`](https://github.com/dojoengine/katana/tree/1.7.0)). So, the next version ie `1.7.0-alpha.1` will be coming from main. Hence, why this changes are needed. None of the options provided by the `release-dispatch.yml` workflow will be able to advance the current version (1.6.3) of the main branch to the desired version (ie `1.7.0-alpha.1`).