ci(test): build binary in ci (#229)
Build the Katana binary as part of the test CI pipeline instead of relying on the [`katana-dev`](https://github.com/dojoengine/katana/pkgs/container/katana-dev) container that we use in the CI runner.

This ensure that when running tests that require using the binary (e.g., `katana-node-bindings` tests), we always run it against the most updated Katana binary that includes all the changes that happen in a PR. Otherwise, the tests would be executed against an outdated Katana as it's not feasible to update the `katana-dev` Docker image on every commit.