feat: bump `blockifier` to `0.15.0-rc.3` (#248)
Bumps `blockifer` to the latest [upstream] tag `v0.15.0-rc.3`. We still rely on our fork, so the version that is used for this bump is a [patched version] of the upstream tag. The patched is basically the branch `blockifier/cairo-2.11` but rebased onto the upstream tag
`v0.15.0-rc.3`.

The `snos-integration-test` will be removed temporarily until we have a version of SNOS using `cairo-lang-*@2.12.0-dev.1` as currently it's causing a version conflict issue with the cairo-lang deps. The test will most likely be added back once the new SNOS implementation is ready. The [current one] is pretty much deprecated and its development has stopped.

[upstream]: https://github.com/starkware-libs/sequencer/releases/tag/v0.15.0-rc.3
[patched version]: https://github.com/dojoengine/sequencer/tree/blockifier/v0.15.0-rc.3
[current one]: https://github.com/keep-starknet-strange/snos