feat(rpc): Starknet RPC extensions  (#206)
Extending the Starknet JSON-RPC server to include several new methods: `starknet_getBlocks`, `starknet_getTransactions`, `starknet_transactionNumber`. These methods are added primarily to support a feature on Explorer (https://github.com/cartridge-gg/explorer/issues/184) - to display a list of blocks and transactions which aren't possible with the existing set of Starknet JSON-RPC (which Explorer solely rely upon atm).

`starknet_getBlocks`, and `starknet_getTransactions` are intended to be the block and transactions equivalent of `starknet_getEvents` - where instead of returning events, they return blocks and transactions repsectively. The methods are meant to return enough information to display the blocks/transactions list no Explorer.

While `starknet_transactionNumber` is the equivalent of `starknet_blockNumber` but for transaction. In this case, the transaction number is a key that is assigned to every transactions and used mainly for indexing purpose in the Katana database. As such its semantic is very much internal to Katana. Considering Katana only stores executed (i.e., mined transactions), the transaction number can also be seen as the order of inclusion relative to all mined transactions ever.

This is intended to be a temporary solution until we've implemented a more efficient solution for querying this type of data from Katana. Ideally, as GraphQL queries with an underlying database that allows for efficient sequential queries.