refactor(controller): remove dependency on `account_sdk` (#260)
The crate is removed because it has the `starknet` crate in its dependency tree. This is part of an iterative effort to reduce the impact of `starknet` crate in the codebase.

The latest version (as of today) of `starknet` crate ie version `0.17.0` bundles 2 (3 if includes MSRV bump) breaking changes - RPC upgrade, and minor bump of `starknet-types-core` to `0.2.0`. The minor bump is the change that caused the most damage as it conflicts with other crates that has `starknet-types-core` in their dependency trees as they are still on the `0.1` minor version. Our reliance on `starknet` crate for the RPC types has initially forced us to use the `0.17.0` version to support the recent Starknet JSON-RPC 0.9.0 [upgrade]. But now with #256 we are moving away from relying on the crate which would prevent us from being too tightly coupled with a specific version of `starknet`, and to use a version of the crate that is compatible with the rest of the related crate ecosystem.

[upgrade]: https://github.com/xJonathanLEI/starknet-rs/pull/778