refactor(rpc-types): define RPC types locally (#256)
Completely remove the reliance on `starknet-rs` to provide RPC types. coincidentally, this would allow us to use an earlier version of `starknet-rs` to avoid `starknet-types-core` version conflicts.

---

Taken from [#260]:

> The latest version of `starknet` crate (ie `0.17.0`) bundles 2 (3 if includes higher MSRV) breaking changes - RPC upgrade, and minor bump of `starknet-types-core` to `0.2.0`. The minor bump is the change that caused the most damage as it conflicts with other crates that has `starknet-types-core` in their dependency trees as they are still on the `0.1` minor version. Our reliance on `starknet` crate for the RPC types has initially forced us to use the `0.17.0` version to support the recent Starknet JSON-RPC 0.9.0 [upgrade].

With this, Katana is no longer dependent on `0.17.0` version of `starknet`, or any specifc version for that matter every time we need to support newer version of the Starknet JSON-RPC specification. Thus, allowing us to use a version of the crate that is compatible with the rest of the related crate ecosystem.

[upgrade]: https://github.com/xJonathanLEI/starknet-rs/pull/778
[#260]: https://github.com/dojoengine/katana/pull/260