Bump `blockifier` to `v0.15.0-rc.2` (#271)
Bumps `blockifier` to upstream tag [`v0.15.0-rc.2`]. This is necessary to make Katana works with the new [SNOS] implementation as it includes some changes in how the transaction fee is computed. SNOS is not compatible with the current version of `blockifier` due to a fee mismatch error.

## Database compatibility

The tag introduces a change in the memory representation of `TransactionExecutionInfo` struct. This breaks compatibility with the current database format.

---

We previously did a [bump] to v0.15.0-rc.3 but it was reverted due to the breaking database change.

[`v0.15.0-rc.2`]:
https://github.com/dojoengine/sequencer/commits/blockifier/v0.15.0-rc.2
[SNOS]: https://github.com/chudkowsky/snos-poc
[bump]: https://github.com/dojoengine/katana/pull/248