refactor(feeder-gateway): clean up types + 0.14.0 support (#272)
This is primarily to update the feeder gateway objects to ensure it is compatible with the recent feeder API [upgrade] of Starknet 0.14.0 and to do some clean up to prevent redundant definitions whenever possible as most of them have similar network format as the RPC types.

The 0.14.0 upgrade also introduces a new endpoint to the feeder gateway, `get_preconfirmed_block`, but I didn't manage to have it working (keep getting not found and bad request errors on both mainnet and sepolia). But, we're not using method anywhere yet so I don't include it in this commit.

[upgrade]:
https://community.starknet.io/t/sn-0-14-0-pre-release-notes/115618#p-2359352-feeder-api-15