feat(feeder-gateway): feeder gateway server (#275)
Adds a server implementation serving the Starknet feeder gateway endpoints that can be used to query the sequencer's state.

This is primarily intended to aid in development of the syncing feature - avoid having to rely entirely on the official feeder gateway offered by Starknet - by imitating the gateway endpoints.

There are very good reasons for implementing this server:

* Everything can be done locally - provided we have synced once (if we want to use Starknet mainnet/sepolia data), or use locally generated data entirely (no syncing is required at all).
* Starknet feeder gateway can rate limit our requests.

## Feeder gateway

The Starknet feeder gateway serves as a gateway for querying the centralized sequencer on the state of Starknet. It was meant to be used as a temporary solution until full nodes that support [Starknet’s JSON-RPC](https://github.com/starkware-libs/starknet-specs#starknet-specifications) mature and are able to meet the needs of the ecosystem. Starknet doesn't yet have a P2P network. Meaning, currently, full nodes sync by fetching data from the feeder gateway.

## Reference(s):

* Changes on the feeder gateway in Starknet 0.14.0 - https://community.starknet.io/t/sn-0-14-0-pre-release-notes/115618#p-2359352-gateway-and-feeder-gateway-api-13
* List of endpoints available after its deprecation in favour of the JSON-RPC specs - https://community.starknet.io/t/feeder-gateway-deprecation/100233