feat(gateway): add `add_transaction` gateway endpoint (#303)
Adds `add_transaction` method of the gateway endpoint for submitting transactions directly to the sequencer mempool. The implementation is mostly based on [Pathfinder's].

As such, it is appropriate to rename the crate to `katana-gateway` instead of `katana-feeder-gateway` - the crate now longer constitutes just feeder gateway related items.

> _Drop the "feeder". Just "gateway". It's cleaner_

## Context

As of now, Starknet doesn't have a P2P network yet. So, the only way nodes currently submit transactions to the network is via the `add_transaction` gateway endpoint.

[Pathfinder's]:
https://github.com/eqlabs/pathfinder/blob/631005f77a4776e5029c0fc7a332bec7377bc7c8/crates/gateway-client/src/lib.rs#L505-L563