feat(sync): return lower bound of stages' last block processed (#316)
This is to allow the pipeline to always execute a stage with a block range that is limited by the configured pipeline batch size - prevents the case where a stage might get executed with an extremely large block range.

Let's give an example of a `Pipeline` with a batch size of 10, and consists of two stages - Stage A, and Stage B - and they each have a checkpoint of **5** and **30** respectively (assuming that we're resuming the pipeline from an existing broken previous run).

If the first `Pipeline::run_to` cycle has an input, `to`, of 10, Stage A will be executed from 5 -> 10 while Stage B gets _skipped_ (because its checkpoint >= 10). Then, the method will return block number **30** because Stage B is the **last** stage to be executed:

```rust
if checkpoint >= to {
    info!(target: "pipeline", %id, "Skipping stage.");

    if i == last_stage_idx {
        return Ok(checkpoint);
    }

    continue;
}
```

On the next iteration of `Pipeline::run_to`, the method will be called with a `to` of **30**. This will trigger an execution of Stage A with starting from block 10 -> 30. Obviously this example uses small numbers for simplicity, but you could imagine if the checkpoint difference between the Stage A and Stage B is very large, Stage A could get
executed with an insanely large range of blocks (eg 0..10,000) and because we don't yet handle partial failure all the progress would be lost if an error occur partway.