feat(gateway): implements `get_public_key` and `get_signature` endpoints (#324)
These endpoints will be necessary for verifying that the synced blocks are sequenced by the right address i.e., the `sequencer_address` field in the block header.

`get_public_key` returns the public key of sequencer associated by the gateway, while `get_signature` returns the signature of the block as signed using the sequencer's public key.

As currently there's only a single centralized entity that acts as the sequencer (i.e., the official sequencer operated by StarkWare), this is how the public key and signature are obtained. Eventually this may change once Starknet employ decentralized sequencing.