fix(storage): store replaced class updates (#328)
Currently, Katana doesn't store the entries of the `StateUpdates.replaced_classes` field. For context, the field is a mapping of class changes that are made using the `replace_class` syscall on an already deployed contract.

We can't store the `replaced_classes` entries directly in the `ClassChangeHistory` table without any modification. The table is currently intended to store class changes that happen due to contract deployments i.e., the entries in `StateUpdates.deployed_contracts`. The table value doesn't distinguish on the event that cause the class change. So, the simplest approach is to add a new field in the value type of the `ClassChangeHistory` table for preserving that information:

```rust
#[derive(Debug, Default, PartialEq, Eq)]
#[cfg_attr(test, derive(::arbitrary::Arbitrary))]
pub struct ContractClassChange {
    /// The type of class change.
    pub r#type: ContractClassChangeType,
    /// The address of the contract whose class has changed.
    pub contract_address: ContractAddress,
    /// The updated class hash of `contract_address`.
    pub class_hash: ClassHash,
}
```