refactor(node): update tip watcher to be subscription-based (#341)
The idea behind this change is to decouple the `ChainTipWatcher` struct from the specific components that need to react to tip changes.

The `ChainTipWatcher` is currently associated with the syncing `Pipeline` - it holds a handle to the pipeline component that it'd use for updating the pipeline's tip. The current design is fine but it would require adding another component's 'handle' to the `ChainTipWatcher` struct for every new component.

I initially started with turning the `ChainTipWatcher` to use 'hooks' that would be executed every time a new tip is received. But that would put the responsibility of running the hooks to `ChainTipWatcher` itself and I think subscription-based design makes the component more lightweight and completely shoves that responsibility away from it. But
this means we need to create a dedicated task for each component that needs to react to chain tip changes.