refactor(rpc): pending block abstraction (#345)
Abstract how pending/pre-confirm block is handled in the Starknet RPC server. This change introduces the `PendingBlockProvider` trait where pending/pre-confirm block source (eg `BlockProducer`) will have to implement in order to support querying the block data from the RPC.

This is mainly done in order to have the Starknet RPC server be integrated with the full node mode, where the source of the pending/pre-confirm block is from the feeder gateway and not the block producer (for sequencer node).