feat: preserve class abi (#356)
the `abi` field in rpc format for the contract class can be an arbitrary string. this is what the jsonrps [spec] says:

```json
   "abi": {
            "title": "ABI",
            "type": "string",
            "description": "The class ABI, as supplied by the user declaring the class"
          }
```

currently, if the rpc abi string can't be converted into the `ContractAbi` type, it will fail during the deserialization phase. the abi string must be preserved as well because it's part of the class hash computation. so we need to preserve the raw string even if it's just some random string.

Starknet class with arbitrary abi string: `0x22b7d06d355ddb1c485ec9cc09b488e77e63318db8833104aed56d598416417` where the abi string returned by the RPC is literally this `"Declared from starknet-rs test case. Timestamp (ms): 1709974545243"`.

superseded #353 .

this change also include a database version bump ( 7 -> 8 ) because of the change made to the value type of the `Classes` table i.e., the `katana_primitives::class::ContractClass` enum.

[spec]:
https://github.com/starkware-libs/starknet-specs/blob/c2e93098b9c2ca0423b7f4d15b201f52f22d8c36/api/starknet_api_openrpc.json#L3215C8-L3219C12