fix(provider): historical provider doesn't respect pinned block number (#360)
The fork version of the `HistoricalStateProvider` should falls back to requesting data from the forked network ONLY when the pinned block number (i.e., the block number the `HistoricalStateProvider` is created for - the `block_id` given to `<ForkedProvider as
StateFactoryProvider>::historical`) is less than or equal to the block number which katana is forked at.

Currently, `HistoricalStateProvider` would immediately falls back to the forked network. The request is then handled by the `Backend` and `Backend` is pinned to the forked block number and not the block number the state provider is for.