feat: add migrated class hashes in state update (#379)
Adds a new field to the state diff types for class hashes whose CASM hash has been migrated to the new hash function. This changes are part of the RPC v0.10.0 and Starknet 0.14.1 update.

## Database Changes

Introduces a new table, `MigratedCompiledClassHash` - a dupsort table for indexing the list of migrated class hashes according to the block number where the migration occurred.

This change doesn't require bumping the database version as it is backward compatible. Queries performed on the table would not fail and simply return an empty result. Which is still the expected behaviour considering that this table only make sense for Katana with 0.14.1 support which versions prior to this PR have no support for.

## References

* Starknet 0.14.1 pre-release notes. https://community.starknet.io/t/starknet-v0-14-1-prerelease-notes/116032
* SNIP34 proposal for CASM hash change. https://community.starknet.io/t/snip-34-more-efficient-casm-hashes/115979