refactor(primitives): remove usages of `short_string!` macro (#389)
This change is part of the effort to remove the dependency on the `starknet` crate. The `short_string!` macro from that crate was previously used to construct `ShortString` values at compile time, but eliminating this dependency requires an alternative approach. The previous implementation using `heapless::String` did not support const construction of arbitrary strings, so it has been replaced with a custom stack-allocated data structure that enables compile-time construction via `ShortString::from_ascii`.